Call stitch pipeline directly from Prefect task
